+++ src/category.py
@@ -1,22 +1,24 @@
-class Category:
-    total_categories = 0
-    total_products = 0
+from src.product import Product
 
-    def __init__(self, name: str, description: str, products=None):
+
+class Category:
+    def __init__(self, name, description, products=None):
         self.name = name
         self.description = description
-        self.products = products if products is not None else []
-        Category.total_categories += 1
-        Category.total_products += len(self.products)
+        self._products = products if products is not None else []
 
     @property
-    def category_count(self):
-        return Category.total_categories
+    def products(self):
+        return [f"{product.name}, {product.price} руб. Остаток: {product.quantity} шт." for product in self._products]
+
+    def add_product(self, product: Product):
+        if isinstance(product, Product):
+            self._products.append(product)
 
     @property
     def product_count(self):
-        return Category.total_products
+        return len(self._products)
 
-    def add_product(self, product):
-        self.products.append(product)
-        Category.total_products += 1
+    def __str__(self):
+        total_quantity = sum(product.quantity for product in self._products)
+        return f"{self.name}, количество продуктов: {total_quantity} шт."
