--- tests/test_str.py
@@ -1,20 +1,0 @@
-from src.category_str import Category
-from src.product_str import Product
-
-
-def test_product_str():
-    product = Product("Samsung Galaxy S23 Ultra", "256GB, Серый цвет, 200MP камера", 180000.0, 5)
-    assert str(product) == "Samsung Galaxy S23 Ultra, 180000.0 руб. Остаток: 5 шт."
-
-
-def test_category_str():
-    product1 = Product("Samsung Galaxy S23 Ultra", "256GB, Серый цвет, 200MP камера", 180000.0, 5)
-    product2 = Product("Iphone 15", "512GB, Gray space", 210000.0, 8)
-    category = Category("Смартфоны", "Описание категории", [product1, product2])
-    assert str(category) == "Смартфоны, количество продуктов: 13 шт."
-
-
-def test_product_add():
-    product1 = Product("Samsung Galaxy S23 Ultra", "256GB, Серый цвет, 200MP камера", 180000.0, 5)
-    product2 = Product("Iphone 15", "512GB, Gray space", 210000.0, 8)
-    assert product1 + product2 == 2700000.0  # 180000*5 + 210000*8
